+++ app/core/entity/Token.ts
@@ -1,14 +1,37 @@
+import dayjs from 'dayjs';
 import { Entity, EntityData } from './Entity';
 import { EasyData, EntityUtil } from '../util/EntityUtil';
 
-interface TokenData extends EntityData {
+export enum TokenType {
+  granular = 'granular',
+  classic = 'classic',
+}
+interface BaseTokenData extends EntityData {
   tokenId: string;
   tokenMark: string;
   tokenKey: string;
-  cidrWhitelist: string[];
+  cidrWhitelist?: string[];
   userId: string;
-  isReadonly: boolean;
-  isAutomation: boolean;
+  isReadonly?: boolean;
+  type?: TokenType;
+}
+
+interface ClassicTokenData extends BaseTokenData{
+  isAutomation?: boolean;
+}
+interface GranularTokenData extends BaseTokenData {
+  name: string;
+  description?: string;
+  allowedScopes?: string[];
+  allowedPackages?: string[];
+  expires: number;
+  expiredAt: Date;
+}
+
+type TokenData = ClassicTokenData | GranularTokenData;
+
+export function isGranularToken(data: TokenData): data is GranularTokenData {
+  return data.type === TokenType.granular;
 }
 
 export class Token extends Entity {
@@ -19,6 +42,13 @@ export class Token extends Entity {
   readonly userId: string;
   readonly isReadonly: boolean;
   readonly isAutomation: boolean;
+  readonly type?: TokenType;
+  readonly name?: string;
+  readonly description?: string;
+  readonly allowedScopes?: string[];
+  readonly expiredAt?: Date;
+  readonly expires?: number;
+  allowedPackages?: string[];
   token?: string;
 
   constructor(data: TokenData) {
@@ -27,13 +57,27 @@ export class Token extends Entity {
     this.tokenId = data.tokenId;
     this.tokenMark = data.tokenMark;
     this.tokenKey = data.tokenKey;
-    this.cidrWhitelist = data.cidrWhitelist;
-    this.isReadonly = data.isReadonly;
-    this.isAutomation = data.isAutomation;
+    this.cidrWhitelist = data.cidrWhitelist || [];
+    this.isReadonly = data.isReadonly || false;
+    this.type = data.type || TokenType.classic;
+
+    if (isGranularToken(data)) {
+      this.name = data.name;
+      this.description = data.description;
+      this.allowedScopes = data.allowedScopes;
+      this.expiredAt = data.expiredAt;
+      this.allowedPackages = data.allowedPackages;
+    } else {
+      this.isAutomation = data.isAutomation || false;
+    }
   }
 
   static create(data: EasyData<TokenData, 'tokenId'>): Token {
     const newData = EntityUtil.defaultData(data, 'tokenId');
+    if (isGranularToken(newData) && !newData.expiredAt) {
+      newData.expiredAt = dayjs(newData.createdAt).add(newData.expires, 'days').toDate();
+    }
     return new Token(newData);
   }
+
 }